Reset de Migraciones - Desactivado Borrado en Cascada

+++ ERPAPI/Controllers/InsuranceEndorsementController.cs
@@ -178,7 +178,7 @@ namespace ERPAPI.Controllers
 
                     await _context.SaveChangesAsync();
 
-                    double sumacreditos = 0, sumadebitos = 0;
+                    decimal sumacreditos = 0, sumadebitos = 0;
                     if (sumacreditos != sumadebitos)
                     {
                         transaction.Rollback();
