+++ ERPAPI/Models/EndososCertificadosLine.cs
@@ -32,15 +32,15 @@ namespace ERPAPI.Models
         public string SubProductName { get; set; }
 
         [Display(Name = "Cantidad")]
-        public double Quantity { get; set; }        
+        public decimal Quantity { get; set; }        
 
         [Display(Name = "Precio")]
-        public double Price { get; set; }
+        public decimal Price { get; set; }
 
         [Display(Name = "Valor endoso")]
-        public double ValorEndoso { get; set; }
+        public decimal ValorEndoso { get; set; }
 
-        public double Saldo { get; set; }
+        public decimal Saldo { get; set; }
 
 
     }
