+++ ERPAPI/Models/GoodsDeliveredLine.cs
@@ -44,14 +44,14 @@ namespace ERPAPI.Models
 
 
         [Display(Name = "Cantidad")]
-        public double Quantity { get; set; }
+        public decimal Quantity { get; set; }
         [Display(Name = "Sacos")]
         public int QuantitySacos { get; set; }
 
         [Display(Name = "Precio")]
         public double Price { get; set; }
         [Display(Name = "Total")]
-        public double Total { get; set; }
+        public decimal Total { get; set; }
         [Display(Name = "Bodega")]
         public Int64 WareHouseId { get; set; }
