+++ ERPAPI/Models/VendorInvoiceLine.cs
@@ -39,21 +39,21 @@ namespace ERPAPI.Models
         [Display(Name = "Descripcion")]
         public string Description { get; set; }
         //[Display(Name = "Cantidad")]
-        //public double Quantity { get; set; }
+        //public decimal Quantity { get; set; }
         //[Display(Name = "Precio")]
-        //public double Price { get; set; }
+        //public decimal Price { get; set; }
         [Display(Name = "Monto")]
-        public double Amount { get; set; }
+        public decimal Amount { get; set; }
         
 
         //[Display(Name = "Porcentaje descuento")]
-        //public double DiscountPercentage { get; set; }
+        //public decimal DiscountPercentage { get; set; }
         //[Display(Name = "Monto descuento")]
-        //public double DiscountAmount { get; set; }
+        //public decimal DiscountAmount { get; set; }
         //[Display(Name = "Subtotal")]
-        //public double SubTotal { get; set; }
+        //public decimal SubTotal { get; set; }
         [Display(Name = "% Impuesto")]
-        public double TaxPercentage { get; set; }
+        public decimal TaxPercentage { get; set; }
         [Display(Name = "Código Impuesto")]
         public string TaxCode { get; set; }
 
@@ -62,8 +62,8 @@ namespace ERPAPI.Models
         public virtual Tax Tax { get; set; }
 
         [Display(Name = "Monto Impuesto")]
-        public double TaxAmount { get; set; }
-        public double Total { get; set; }
+        public decimal TaxAmount { get; set; }
+        public decimal Total { get; set; }
 
         public long AccountId { get; set; }
